dev-texlive/texlive-langjapanese: only remove man page if USE=doc
Closes: https://bugs.gentoo.org/946051

+++ dev-texlive/texlive-langjapanese/texlive-langjapanese-2024_p72817.ebuild
@@ -181,6 +181,9 @@ src_prepare() {
 	# e(u)ptex are installed by texlive-core[cjk]
 	sed -i '/AddFormat name=eptex /d' tlpkg/tlpobj/ptex.tlpobj || die
 	sed -i '/AddFormat name=euptex /d' tlpkg/tlpobj/uptex.tlpobj || die
-	# ptekf.1 is installed by dev-libs/ptexenc
-	rm texmf-dist/doc/man/man1/ptekf.1 || die
+
+	if use doc; then
+		# ptekf.1 is installed by dev-libs/ptexenc
+		rm texmf-dist/doc/man/man1/ptekf.1 || die
+	fi
 }
